+++ src/agentcage/backends/container.py
@@ -42,7 +42,6 @@ class ContainerBackend:
             "agentcage-proxy",
             os.path.join(containers_dir, "Containerfile.proxy"),
             build_context,
-            no_cache=True,
             cap_add=["CAP_CHOWN", "CAP_FOWNER", "CAP_SETUID", "CAP_SETGID", "CAP_DAC_OVERRIDE"],
             quiet=quiet,
         )